Add Interchain Router service
This change adds the skeleton of an Interchain Router that exposes
a gRPC service for retrieving an external preimage.

It creates the basic proto file for the ExternalPreimageService,
and adds a placeholder function for retrieving the preimage, which
currently throws an error as it is unimplemented.

It also initializes the Interchain Router on BrokerDaemon startup,
using the same pattern established by `GrpcServer`.