Store block order id directly on fills and orders
This change gives responsibility for tracking which block order
a fill/order belongs to directly to that fill/order.

Rather than relying on sublevels prefixed by block order id,
we instead have a flat hierarchy with all fills and orders comingled.
Each fill and order stores their block order as a prefix to their id,
allowing the block order worker to retrieve them with a ranged query.

This should have no impact on performance, since it's the same
operation that sublevel was performing under the hood, but it gives
us more flexibility to apply indexes to the fill and order stores.