InterchainRouter: Check that order for preimage request matches our expectation
When the Interchain Router receives a request for a preimage, it looks up the
order associated with the payment hash and ensures that it matches the request.

If we already settled the order (and therefore already have the preimage), we
return it immediately to avoid paying twice.

Specifically, it checks whether the currency and amount are sufficient and that
the current block height is not higher than the extended time lock.

It also checks that we actually have an order for the hash, but not too many.

To do this, we expose the ordersByHash index to the InterchainRouter during
BrokerDaemon set up.

This still leaves undone:
- checking that the timelock is sufficient for a downstream payment (this
may get feedback by actually attempting the payment rather than shortcutting it
at the router)
- making the downstream payment
- returning the preimage to the requesting node when requested downstream