Translate swaps using the Interchain Router
This change exposes the engines to the Interchain Router, selects
the correct engine when handling an inbound request for a preimage
and making a payment to that engine to retrieve the preimage from
the destination chain/network.

It depends on a forthcoming change to LndEngine to expose the
`translateSwap` action.

It solidifies the design choice of Interchain Router being standalone
and not modifying the state of the rest of the application, or
interacting with it in any way beyond reading Orders from the data store,
which are acting as our Routing Entries.