Complete order by retrieving preimage and sending to the Relayer
This change adds a new transition to the order state machine
that retrieves the preimage from the engine for the settled swap
(which may be a long running operation) and then provides that
preimage to relayer, proving that the swap has completed, and
triggering the relayer to release the Maker and Taker deposits.

It is designed to restart when being re-hydrated from storage
although we don't currently have support for re-starting orders
when the daemon restarts.

This change depends on the getSettledSwapPreimage method added
to the LndEngine which in turn relies on the change in LND to
persist external preimages locally once they are settled.